Invierte

+++ operaciones_de_matrices_dinamicas_float_diagonalizable/Include/operaciones.h
@@ -12,7 +12,7 @@ void matrix_save(FILE *stream, matrix mat);
 int matrix_get(FILE *stream, matrix *mat);
 int add(matrix mat1, matrix mat2, matrix *res);
 int transpose(matrix mat, matrix *res);
-int scalar_mult(matrix mat, int scalar, matrix *res);
+int scalar_mult(matrix mat, float scalar, matrix *res);
 int matrix_mult(matrix mat1, matrix mat2, matrix *res);
 void matrix_save_setpretty(void);
 int matrix_alloc(matrix *mat);
